useLocalAwareness to track cursor position.
Also useRemoteAwareness to track peers.

Trying to reset cursor position after remote update, my best idea is to
use the DocHandle to listen for changes, and record the lastUserEdited
in the doc itself.  If the lastUserEdited doesn't match the current
user, then update cursor position, otherwise leave it.  Seems like
something that should be left into the emphemeral storage in the
DocHandler itself (and exposed with local awareness / remote awareness).

Changing the cursor position at every update means updating it locally
moves the cursor around to the last position, breaking the editor.