fixing cursor movement on remote edit.
Ignore the last comment message.
Rather, to see what was going on with the cursor, notice how
@monaco-editor/react updates the value of the Editor's model, see
https://github.com/suren-atoyan/monaco-react/blob/master/src/Editor/Editor.tsx#L106

It has a `forceMoveMarkets: true`, but since its using the full range of
the model, its going to force markers to move to the end.  A quick hack
to update manually (watching the handle for updates rather than
useDocument and value) fixes the cursor for several updates, though
undo/redo is now weird (probably can fix that just be showing the stack
of doc changes).

A better idea might be to compute the Automerge doc changes, and then
apply monaco edits for each individual change, rather than nuking the
whole thing.  Maybe in next commit.