using DocHandle's events rather than update func.
I think I misread the code initially, and the events are a better way to
do this.  update was for something else.

I could use the `ephemeral-message` event to update peerStates manually
instead of the useRemoteAwareness hook, but I'm running out of time to
wrap this up.  So we'll leave it for now.

Performance is much better when editing large files, except that it did
crash chomium when I pasted in the large file from the firefox client.